Prep for removal of aws-cognito-promises.
Temp copy the aws-cognito-promises package into this one. I will refactor the
code to live within the sagas instead.

This fixes the auth signup for omt-tech by adding locale and phoneNumber fields.
Note that we will later update the call signature to allow for an array of
attributes for more flexiblility.